clean up path for new chart component
