+++ app/scripts/components/common/chart/alt-title.tsx
