+++ app/scripts/components/common/chart/block.tsx
