+++ app/scripts/components/common/chart/constant.js
