+++ app/scripts/components/common/chart/index.tsx
