+++ app/scripts/components/common/chart/legend.tsx
